+++ lib/vagrant-skytap/config.rb
@@ -75,6 +75,21 @@ module VagrantPlugins
       # @return [String]
       attr_accessor :guestos
 
+      # The VMware guest OS setting for this machine.
+      #
+      # @return [Array]
+      attr_accessor :disks
+
+      # The Skytap Environment Name.
+      #
+      # @return [String]
+      attr_accessor :environment_name
+
+      # The Skytap Project to assign Environment.
+      #
+      # @return [String]
+      attr_accessor :project_id
+
       def initialize(region_specific=false)
         @username               = UNSET_VALUE
         @api_token              = UNSET_VALUE
@@ -83,11 +98,14 @@ module VagrantPlugins
         @vpn_url                = UNSET_VALUE
         @instance_ready_timeout = UNSET_VALUE
         @region                 = UNSET_VALUE
+        @environment_name       = UNSET_VALUE
+        @project_id             = UNSET_VALUE
 
         @cpus                   = UNSET_VALUE
         @cpuspersocket          = UNSET_VALUE
         @ram                    = UNSET_VALUE
         @guestos                = UNSET_VALUE
+        @disks                  = UNSET_VALUE
       end
 
       #-------------------------------------------------------------------
@@ -112,12 +130,17 @@ module VagrantPlugins
         # Set the default timeout for runstate changes (e.g. running a VM)
         @instance_ready_timeout = 300 if @instance_ready_timeout == UNSET_VALUE
 
+        # Environment Name and Project ID default to nil
+        @environment_name = nil if @environment_name == UNSET_VALUE
+        @project_id = nil if @project_id == UNSET_VALUE
+
         # Hardware settings default to nil (will be obtained
         # from the source VM)
         @cpus          = nil if @cpus          == UNSET_VALUE
         @cpuspersocket = nil if @cpuspersocket == UNSET_VALUE
         @ram           = nil if @ram           == UNSET_VALUE
         @guestos       = nil if @guestos       == UNSET_VALUE
+        @disks         = nil if @disks         == UNSET_VALUE
 
         # Mark that we finalized
         @__finalized = true